added shared controller and template
