+++ src/routes/routes.ts
@@ -39,6 +39,7 @@ import * as updateDirectorName from "../controllers/update/update.director.name.
 import * as updateDirectorNationality from "../controllers/update/update.director.nationality.controller";
 import { isFeatureEnabled } from "../middleware/is.feature.enabled.middleware";
 import * as updateDirectorCorrespondenceAddress from "../controllers/update/update.director.correspondence.address.controller";
+import * as updateDirectorResidentialAddressManual from "../controllers/update/update.director.residential.address.manual.controller";
 import * as urls from "../types/page.urls";
 import { AP01_ACTIVE, CH01_ACTIVE } from "../utils/properties";
 import { checkYourAnswersMiddleware } from "../middleware/check.your.answers.middleware";
@@ -170,3 +171,5 @@ router.post(urls.UPDATE_DIRECTOR_NATIONALITY, companyAuthenticationMiddleware, n
 router.get(urls.UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS, isFeatureEnabled(CH01_ACTIVE), companyAuthenticationMiddleware, checkYourAnswersMiddleware(), updateDirectorCorrespondenceAddress.get);
 router.post(urls.UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS, companyAuthenticationMiddleware, updateDirectorCorrespondenceAddress.post);
 
+router.get(urls.UPDATE_DIRECTOR_RESIDENTIAL_ADDRESS_MANUAL, isFeatureEnabled(CH01_ACTIVE), companyAuthenticationMiddleware, updateDirectorResidentialAddressManual.get);
+router.post(urls.UPDATE_DIRECTOR_RESIDENTIAL_ADDRESS_MANUAL, companyAuthenticationMiddleware, updateDirectorResidentialAddressManual.post);
