+++ src/types/template.paths.ts
@@ -40,5 +40,6 @@ export enum Templates {
   UPDATE_DIRECTOR_OCCUPATION = "update/update-director-occupation",
   UPDATE_DIRECTOR_SUBMITTED = "update/update-director-submitted",
   UPDATE_DIRECTOR_NATIONALITY = "update/update-director-nationailty",
-  UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS = "update/update-director-correspondence-address"
+  UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS = "update/update-director-correspondence-address",
+  UPDATE_DIRECTOR_RESIDENTIAL_ADDRESS_MANUAL = "update/update-director-residential-address-manual",
 }
