+++ views/includes/page-templates/director-residential-address-manual.html
@@ -1,0 +1,115 @@
+{% extends "layout.html" %}
+
+{% set title = "Enter the director's home address " %}
+
+{% block pageTitle %}
+  {% include "includes/page-title.html" %}
+{% endblock %}
+
+{% block backLink %}
+  {% if residential_address_back_param === "confirm-residential-address" %}
+    {% set backLinkText = "Go back to 'Confirm where the director lives'" %}
+  {%else%}
+    {% set backLinkText = "Go back to 'Find the director's home address'" %}
+  {%endif%}
+  {% include "includes/back-link.html" %}
+{% endblock %}
+
+{% block content %}
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds-from-desktop">
+
+      {% include "includes/list/errors.html" %}
+      
+      <form action="" method="post">
+
+        <span class="govuk-caption-l">{{directorName}}</span>
+        <h1 class="govuk-heading-l">Enter the director's home address</h1>
+
+        <div id="residential-property-address-hint" class="govuk-hint">
+          Tell us the director's home address. Only give a business address if they also live there.
+        </div>
+
+        {{ govukInput({
+          errorMessage: errors.residential_address_premises if errors,
+          label: {
+            html: 'Property name or number'
+          },
+          hint: {
+            text: "For example, 'The Mill', '116' or 'Flat 37a'"
+          },
+          id: "residential_address_premises",
+          name: "residential_address_premises",
+          value: residential_address_premises
+        }) }}
+
+        {{ govukInput({
+          errorMessage: errors.residential_address_line_1 if errors,
+          label: {
+            text: "Address line 1"
+          },
+          id: "residential_address_line_1",
+          name: "residential_address_line_1",
+          value: residential_address_line_1
+        }) }}
+
+        {{ govukInput({
+          errorMessage: errors.residential_address_line_2 if errors,
+          label: {
+            text: "Address line 2 (optional)"
+          },
+          id: "residential_address_line_2",
+          name: "residential_address_line_2",
+          value: residential_address_line_2
+        }) }}
+
+        {{ govukInput({
+          errorMessage: errors.residential_address_city if errors,
+          label: {
+            text: "City or town"
+          },
+          id: "residential_address_city",
+          name: "residential_address_city",
+          value: residential_address_city
+        }) }}
+
+        {{ govukInput({
+          errorMessage: errors.residential_address_county if errors,
+          label: {
+            text: "County, state, province or region (optional)"
+          },
+          id: "residential_address_county",
+          name: "residential_address_county",
+          value: residential_address_county
+        }) }}
+
+        {% include "includes/inputs/fields/country-input.html" %}
+
+        {{ govukInput({
+          errorMessage: errors.residential_address_postcode if errors,
+          label: {
+            text: "Postcode or ZIP"
+          },
+          classes: "govuk-input govuk-input--width-10",
+          id: "residential_address_postcode",
+          name: "residential_address_postcode",
+          value: residential_address_postcode
+        }) }}
+
+        <div class="govuk-inset-text">
+          <h2 class="govuk-heading-m">{{ publicRegisterInformation }}</h2>
+            <p>We will not show this address on the public Companies House register, unless it’s also used for the company's registered office address or the director’s correspondence address.</p>
+        </div>
+
+
+        {{ govukButton({
+          attributes: {
+            "data-event-id": "continue-button"
+          },
+          text: "Continue"
+        }) }}
+
+      </form>
+    </div>
+  </div>
+{% endblock %}
